Fixed gldownload element in transport
It was being added to the pipeline but not linked because it does not
swallow encoded streams at it is the case with the Pi. I added support
to now if the stream is already encoded and it will not add the element
only if that is the case.

I reworked the whole method to make it more general so you can add
elements and link and sync them at the end.